legacy/requirements.txt
Flask
flask_jwt_extended
requests
urllib3
aiohttp
websocket-client
